polish timings to ensure trigger check results are delivered to trigger

+++ nuget-server/src/jetbrains/buildServer/nuget/server/trigger/impl/PackageCheckerSettings.java
@@ -8,7 +8,7 @@ import org.jetbrains.annotations.NotNull;
  */
 public interface PackageCheckerSettings {
 
-  int getTriggerPollInterval();
+  long getTriggerPollInterval();
 
   long getPackageCheckInterval();
